+++ sql/schema.sql
@@ -53,6 +53,7 @@ CREATE TABLE transactions(
   creation_date DATETIME NOT NULL,
   delivered_date DATETIME,
   expected_date DATETIME,
+  order_status VARCHAR(9) NOT NULL,
   total DECIMAL(10, 2) NOT NULL,
   FOREIGN KEY (user_id) REFERENCES users(id)
     ON DELETE CASCADE